Restructured the Neovim plugin files
