+++ rascal-neovim-plugin/lua/rascal/init.lua
@@ -1,0 +1,3 @@
+local M = {}
+
+return M
